introduce ValueStackRef abstraction
This yielded significant performance improvemetns across the board.
Especially compute intense workloads displayed roughly 15-25% improvements in performance.
Host calls performance seemed to have regressed slightly.
