+++ wasmi_v1/src/engine/stack/mod.rs
@@ -3,7 +3,7 @@ mod values;
 
 pub use self::{
     frames::{CallStack, FuncFrame},
-    values::ValueStack,
+    values::{ValueStack, ValueStackRef},
 };
 use super::{
     code_map::{CodeMap, InstructionsRef},
